requirements.txt
# Existing dependencies
google==3.0.0
google-cloud-core==2.4.1
google-cloud-aiplatform==1.99.0
streamlit==1.28.0
gunicorn==20.0.4
python-dotenv==1.0.0

# Data validation and models
pydantic==2.5.3

# Document processing
PyPDF2==3.0.1
pydicom==2.4.4
pillow==10.1.0

# Configuration
pyyaml==6.0.1

# Async support
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Optional: Advanced NLP (uncomment if needed)
# spacy==3.7.2
# scispacy==0.5.3

# Optional: Database support (uncomment if needed)
# sqlalchemy==2.0.23
# redis==5.0.1